+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 ### added
 - after using `../script suspend`, unsuspend it before time is up using `../script unsuspend`
 ### fixed
-- `../script suspend` is guaranteed to work right and not buffer commands or some other unexpected behaivour
+- `../script suspend` is guaranteed to work right and not buffer commands or some other unexpected behaviour
+- when checking for the presence of moderators in mod_rota, ignore the account the script is running as, since it will always be present.
 
 ## [3.32.1] - 2021-04-15
 ### added