+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [3.33.0] - 2021-04-15
+### added
+- after using `../script suspend`, unsuspend it before time is up using `../script unsuspend`
+### fixed
+- `../script suspend` is guaranteed to work right and not buffer commands or some other unexpected behaviour
+- when checking for the presence of moderators in mod_rota, ignore the account the script is running as, since it will always be present.
+
+## [3.32.1] - 2021-04-15
+### added
+- warn if not running as admin
+- send a warning webhook to the moderators when an account is ignored
+
 ## [3.32.0] - 2021-04-06 - 'Season 3 Day One Patch'
 ### added
 - back and forwards commands